--- a/cocos/2d/CCNode.cpp
+++ b/cocos/2d/CCNode.cpp
@@ -58,7 +58,7 @@ NS_CC_BEGIN
 bool nodeComparisonLess(Node* n1, Node* n2)
 {
 #if CC_64BITS
-    return (n1->_localZOrder < n2->_localZOrder);
+    return (n1->_localZOrderAndArrival < n2->_localZOrderAndArrival);
 #else
     return n1->_localZOrder < n2->_localZOrder || (n1->_localZOrder == n2->_localZOrder && n1->_orderOfArrival < n2->_orderOfArrival);
 #endif
@@ -91,10 +91,9 @@ Node::Node()
 , _transformUpdated(true)
 // children (lazy allocs)
 // lazy alloc
+, _localZOrderAndArrival(0)
 , _localZOrder(0)
-#if !CC_64BITS
 , _orderOfArrival(0)
-#endif
 , _globalZOrder(0)
 , _parent(nullptr)
 // "whole screen" objects. like Scenes and Layers, should set _ignoreAnchorPointForPosition to true
@@ -280,20 +279,15 @@ void Node::setLocalZOrder(int z)
 /// used internally to alter the zOrder variable. DON'T call this method manually
 void Node::_setLocalZOrder(int z)
 {
-#if CC_64BITS
-    _localZOrder = (static_cast<std::int64_t>(z) << 32) | (_localZOrder & 0xffffffff);
-#else
+    _localZOrderAndArrival = (static_cast<std::int64_t>(z) << 32) | (_localZOrderAndArrival & 0xffffffff);
     _localZOrder = z;
-#endif
 }
 
 void Node::updateOrderOfArrival()
 {
-#if CC_64BITS
-    _localZOrder = (_localZOrder & 0xffffffff00000000) | (s_globalOrderOfArrival++);
-#else
-    _orderOfArrival = s_globalOrderOfArrival++;
-#endif
+    auto newArrival = ++s_globalOrderOfArrival;
+    _localZOrderAndArrival = (_localZOrder & 0xffffffff00000000) | (newArrival);
+    _orderOfArrival = newArrival;
 }
 
 void Node::setGlobalZOrder(float globalZOrder)