--- a/cocos/2d/CCNode.h
+++ b/cocos/2d/CCNode.h
@@ -189,7 +189,7 @@ public:
      *
      * @return The local (relative to its siblings) Z order.
      */
-    virtual int getLocalZOrder() const { return _localZOrder.detail.z; }
+    virtual int getLocalZOrder() const { return static_cast<int>(_localZOrder >> 32); }
     CC_DEPRECATED_ATTRIBUTE virtual int getZOrder() const { return getLocalZOrder(); }
 
     /**
@@ -1895,16 +1895,10 @@ protected:
     mutable bool _additionalTransformDirty; ///< transform dirty ?
     bool _transformUpdated;         ///< Whether or not the Transform object was updated since the last frame
 
-    union {
-        struct {
-            std::int32_t z; // The original localZOrder
-            std::uint32_t a; // Order Of Arrival, for avoid sort problem with unstable_sort algorithm.
-        } detail;
-        std::int64_t value; // The value to be used in sort
-    } _localZOrder;               ///< Local order (relative to its siblings) used to sort the node
+    std::int64_t _localZOrder;               ///< Local order (relative to its siblings) used to sort the node
     float _globalZOrder;            ///< Global order used to sort the node
 
-    static unsigned int s_globalOrderOfArrival;
+    static std::uint32_t s_globalOrderOfArrival;
 
     Vector<Node*> _children;        ///< array of children nodes
     Node *_parent;                  ///< weak reference to parent node