--- a/cocos/2d/CCNode.h
+++ b/cocos/2d/CCNode.h
@@ -43,6 +43,12 @@
 #include "physics/CCPhysicsBody.h"
 #endif
 
+#if defined(_M_X64) || defined(_WIN64) || defined(__LP64__) || defined(_LP64) || defined(__x86_64)
+#define CC_64BITS 1
+#else
+#define CC_64BITS 0
+#endif
+
 NS_CC_BEGIN
 
 class GridBase;
@@ -189,7 +195,12 @@ public:
      *
      * @return The local (relative to its siblings) Z order.
      */
+#if CC_64BITS
     virtual int getLocalZOrder() const { return static_cast<int>(_localZOrder >> 32); }
+#else
+    virtual int getLocalZOrder() const { return _localZOrder; }
+#endif
+
     CC_DEPRECATED_ATTRIBUTE virtual int getZOrder() const { return getLocalZOrder(); }
 
     /**
@@ -1895,10 +1906,15 @@ protected:
     mutable bool _additionalTransformDirty; ///< transform dirty ?
     bool _transformUpdated;         ///< Whether or not the Transform object was updated since the last frame
 
+#if CC_64BITS
     std::int64_t _localZOrder;               ///< Local order (relative to its siblings) used to sort the node
+#else
+    int _localZOrder; /// < Local order (relative to its siblings) used to sort the node
+    unsigned int _orderOfArrival; /// < used to preserve sequence while sorting children with the same localZOrder
+#endif
     float _globalZOrder;            ///< Global order used to sort the node
 
-    static std::uint32_t s_globalOrderOfArrival;
+    static unsigned int s_globalOrderOfArrival;
 
     Vector<Node*> _children;        ///< array of children nodes
     Node *_parent;                  ///< weak reference to parent node