… cocos/2d/CCSprite.cpp
… cocos/2d/CCSpriteBatchNode.cpp
… cocos/base/ccMacros.h
… cocos/editor-support/cocostudio/ActionTimeline/CCBoneNode.cpp
… cocos/editor-support/cocostudio/ActionTimeline/CCSkeletonNode.cpp
… cocos/editor-support/cocostudio/CCBone.cpp
… tests/cpp-empty-test/proj.android/jni/main.cpp
… tools/tolua/cocos2dx.ini