* And bring _orderOfArrival back for 32bits program,
* From vs2015 test result, the performance of lambda expression is better than the function: nodeComparisonLess

test program:
```
#include <iostream>
#include <stdio.h>
#include <stdlib.h>
#include <algorithm>
#include <vector>
#include <time.h>
#include <assert.h>

#if 0 // for android
#define  LOG_TAG    "cpp-empty-test"
#define  LOGD(...)  __android_log_print(ANDROID_LOG_DEBUG,LOG_TAG,__VA_ARGS__)
#define logfunc(format,...) __android_log_print(ANDROID_LOG_DEBUG,LOG_TAG,(format "\n"),##__VA_ARGS__)
#else
#define logfunc(format,...) printf((format "\n"),##__VA_ARGS__)
#endif

class SimpleNode
{
public:
    SimpleNode(int z, unsigned int a)
    {
        _localZOrder = 0;
        setLocalZOrder(z);
        setOrderOfArrival(a);
    }

    int getLocalZOrder() const
    {
        return static_cast<int>(((_localZOrder >> 32) & 0xffffffff));
    }

    unsigned int getOrderOfArrival() const
    {
        return static_cast<unsigned int>((_localZOrder) & 0xffffffff);
    }

    void setLocalZOrder(int z)
    {
        _localZOrder = ((0ULL + z) << 32) | (_localZOrder & 0xffffffff);
        _o_localZOrder = z;
    }

    void setOrderOfArrival(unsigned int orderOfArrival)
    {
        _localZOrder = (_localZOrder & 0xffffffff00000000) | orderOfArrival;
        _o_orderOfArrival = orderOfArrival;
    }

    int _o_localZOrder;
    unsigned int _o_orderOfArrival;

    std::int64_t _localZOrder;
};

inline bool nodeComparisonLess64(SimpleNode* n1, SimpleNode* n2)
{
// #if defined(_M_X64) || defined(_WIN64) || defined(__LP64__) || defined(_LP64) || defined(__x86_64)
    return (n1->_localZOrder < n2->_localZOrder);
// #endif
}

void runPerfromanceFunc()
{
    SimpleNode n1(0, 1);
    SimpleNode n2(1, 0);
    SimpleNode n3(-3, 500);


    assert(n3.getLocalZOrder() == -3);
    assert(n3.getOrderOfArrival() == 500);
    assert(n1._localZOrder < n2._localZOrder);
    assert(n3._localZOrder < n1._localZOrder);

    std::vector<SimpleNode*> vec2;
    logfunc("test program is %d bits!", sizeof(void*) << 3);

    for (auto i = 0; i < 1000000; ++i)
        vec2.push_back(new SimpleNode(2, i + 1));

    clock_t start = 0;

    for (int i = 0; i < 10; ++i) {
        start = clock();
        std::stable_sort(vec2.begin(), vec2.end(), [](const SimpleNode* n1, const SimpleNode* n2) {
            return n1->_o_localZOrder < n2->_o_localZOrder;
        });
        logfunc("std::stable_sort, test round:%d, %lf seconds used.", i + 1, (clock() - start) / (double)CLOCKS_PER_SEC);
    }

    for (int i = 0; i < 10; ++i) {
        start = clock();
        std::sort(vec2.begin(), vec2.end(), [](const SimpleNode* n1, const SimpleNode* n2) {
            return n1->_o_localZOrder < n2->_o_localZOrder || (n1->_o_localZOrder == n2->_o_localZOrder && n1->_o_orderOfArrival < n2->_o_orderOfArrival);
        });
        logfunc("std::sort, no optimize orderOfArrival storage, original implement, test round:%d, %lf seconds used.", i + 1, (clock() - start) / (double)CLOCKS_PER_SEC);
    }

    for (int i = 0; i < 10; ++i) {
        start = clock();
        std::sort(vec2.begin(), vec2.end(), [](const SimpleNode* n1, const SimpleNode* n2) {
            return (n1->_localZOrder < n2->_localZOrder);
        });
        logfunc("std::sort(lambda expression), optimize orderOfArrival storage, test round:%d, %lf seconds used.", i + 1, (clock() - start) / (double)CLOCKS_PER_SEC);
    }
    for (int i = 0; i < 10; ++i) {
        start = clock();
        std::sort(vec2.begin(), vec2.end(), nodeComparisonLess64);
        logfunc("std::sort(inline function), optimize orderOfArrival storage, test round:%d, %lf seconds used.", i + 1, (clock() - start) / (double)CLOCKS_PER_SEC);
    }
}

int main(int, char**)
{
    runPerfromanceFunc();

#if defined(_WIN32)
    getchar();
#endif
    return 0;
}

```
32bits test result:
```
test program is 32 bits!
std::stable_sort, test round:1, 0.033000 seconds used.
std::stable_sort, test round:2, 0.034000 seconds used.
std::stable_sort, test round:3, 0.034000 seconds used.
std::stable_sort, test round:4, 0.034000 seconds used.
std::stable_sort, test round:5, 0.034000 seconds used.
std::stable_sort, test round:6, 0.034000 seconds used.
std::stable_sort, test round:7, 0.034000 seconds used.
std::stable_sort, test round:8, 0.034000 seconds used.
std::stable_sort, test round:9, 0.034000 seconds used.
std::stable_sort, test round:10, 0.035000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:1, 0.027000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:2, 0.026000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:3, 0.026000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:4, 0.027000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:5, 0.026000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:6, 0.026000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:7, 0.025000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:8, 0.025000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:9, 0.026000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:10, 0.025000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:1, 0.033000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:2, 0.033000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:3, 0.033000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:4, 0.034000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:5, 0.032000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:6, 0.034000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:7, 0.037000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:8, 0.035000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:9, 0.033000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:10, 0.034000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:1, 0.048000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:2, 0.048000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:3, 0.048000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:4, 0.048000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:5, 0.048000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:6, 0.047000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:7, 0.047000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:8, 0.047000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:9, 0.048000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:10, 0.047000 seconds used.
```
64bits test result:
```
test program is 64 bits!
std::stable_sort, test round:1, 0.054000 seconds used.
std::stable_sort, test round:2, 0.055000 seconds used.
std::stable_sort, test round:3, 0.055000 seconds used.
std::stable_sort, test round:4, 0.054000 seconds used.
std::stable_sort, test round:5, 0.054000 seconds used.
std::stable_sort, test round:6, 0.054000 seconds used.
std::stable_sort, test round:7, 0.054000 seconds used.
std::stable_sort, test round:8, 0.056000 seconds used.
std::stable_sort, test round:9, 0.053000 seconds used.
std::stable_sort, test round:10, 0.058000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:1, 0.032000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:2, 0.032000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:3, 0.032000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:4, 0.031000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:5, 0.033000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:6, 0.031000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:7, 0.031000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:8, 0.031000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:9, 0.032000 seconds used.
std::sort, no optimize orderOfArrival storage, original implement, test round:10, 0.032000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:1, 0.025000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:2, 0.023000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:3, 0.025000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:4, 0.024000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:5, 0.026000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:6, 0.024000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:7, 0.025000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:8, 0.025000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:9, 0.024000 seconds used.
std::sort(lambda expression), optimize orderOfArrival storage, test round:10, 0.025000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:1, 0.044000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:2, 0.047000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:3, 0.043000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:4, 0.043000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:5, 0.044000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:6, 0.044000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:7, 0.044000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:8, 0.043000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:9, 0.044000 seconds used.
std::sort(inline function), optimize orderOfArrival storage, test round:10, 0.045000 seconds used.
```