@halx99 2016-07-29
Xiaomi4 test result(gcc4.9, ndkr10e release mode)
```
D	07-29 15:12:21.595	15188	15188	 cpp-empty-test	cocos_android_app_init
D	07-29 15:12:21.603	15188	15188	 cpp-empty-test	test program is 32 bits!
D	07-29 15:12:22.131	15188	15188	 cpp-empty-test	std::stable_sort, test round:1, 0.172350 seconds used.
D	07-29 15:12:22.303	15188	15188	 cpp-empty-test	std::stable_sort, test round:2, 0.167515 seconds used.
D	07-29 15:12:22.476	15188	15188	 cpp-empty-test	std::stable_sort, test round:3, 0.163518 seconds used.
D	07-29 15:12:22.669	15188	15188	 cpp-empty-test	std::stable_sort, test round:4, 0.173520 seconds used.
D	07-29 15:12:22.825	15188	15188	 cpp-empty-test	std::stable_sort, test round:5, 0.154951 seconds used.
D	07-29 15:12:22.979	15188	15188	 cpp-empty-test	std::stable_sort, test round:6, 0.153649 seconds used.
D	07-29 15:12:23.142	15188	15188	 cpp-empty-test	std::stable_sort, test round:7, 0.155475 seconds used.
D	07-29 15:12:23.299	15188	15188	 cpp-empty-test	std::stable_sort, test round:8, 0.155911 seconds used.
D	07-29 15:12:23.477	15188	15188	 cpp-empty-test	std::stable_sort, test round:9, 0.176782 seconds used.
D	07-29 15:12:23.675	15188	15188	 cpp-empty-test	std::stable_sort, test round:10, 0.193905 seconds used.
D	07-29 15:12:23.999	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:1, 0.298132 seconds used.
D	07-29 15:12:24.285	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:2, 0.282682 seconds used.
D	07-29 15:12:24.598	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:3, 0.304314 seconds used.
D	07-29 15:12:24.998	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:4, 0.368294 seconds used.
D	07-29 15:12:25.504	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:5, 0.470537 seconds used.
D	07-29 15:12:25.794	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:6, 0.285361 seconds used.
D	07-29 15:12:26.197	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:7, 0.400462 seconds used.
D	07-29 15:12:26.623	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:8, 0.422372 seconds used.
D	07-29 15:12:26.985	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:9, 0.360316 seconds used.
D	07-29 15:12:27.355	15188	15188	 cpp-empty-test	std::sort, no optimize orderOfArrival storage, original implement, test round:10, 0.365021 seconds used.
D	07-29 15:12:27.704	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:1, 0.347516 seconds used.
D	07-29 15:12:28.057	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:2, 0.351746 seconds used.
D	07-29 15:12:28.418	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:3, 0.353969 seconds used.
D	07-29 15:12:28.767	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:4, 0.347423 seconds used.
D	07-29 15:12:29.115	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:5, 0.345673 seconds used.
D	07-29 15:12:29.474	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:6, 0.355570 seconds used.
D	07-29 15:12:29.824	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:7, 0.345529 seconds used.
D	07-29 15:12:30.233	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:8, 0.350313 seconds used.
D	07-29 15:12:30.611	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:9, 0.367260 seconds used.
D	07-29 15:12:30.964	15188	15188	 cpp-empty-test	std::sort(lambda expression), optimize orderOfArrival storage, test round:10, 0.352153 seconds used.
D	07-29 15:12:31.432	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:1, 0.451316 seconds used.
D	07-29 15:12:31.888	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:2, 0.391819 seconds used.
D	07-29 15:12:32.338	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:3, 0.432703 seconds used.
D	07-29 15:12:32.792	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:4, 0.447708 seconds used.
D	07-29 15:12:33.261	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:5, 0.452910 seconds used.
D	07-29 15:12:33.724	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:6, 0.454564 seconds used.
D	07-29 15:12:34.192	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:7, 0.461060 seconds used.
D	07-29 15:12:34.658	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:8, 0.459522 seconds used.
D	07-29 15:12:35.123	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:9, 0.450755 seconds used.
D	07-29 15:12:35.589	15188	15188	 cpp-empty-test	std::sort(inline function), optimize orderOfArrival storage, test round:10, 0.457470 seconds used.

```