@minggo 2016-07-29
So the optimization is slower?