@halx99 2016-07-29
because the optimize is for 64bits