@halx99 2016-07-29
Our corp have no 64bits android device to test the result, could compile a 64bits program and test it?