@minggo 2016-07-29
But it slows down 32bit. It is not acceptable.