@halx99 2016-07-29
You know, my PR implement use CC_64BITS to choice the compretion variable