@minggo 2016-07-29
Your test case doesn't use macro to test? Why not share your test case based on `cpp-empty-test`?