@halx99 2016-07-29
According to win64 the optimize is fastest:
![nya689 f64el5vxz4n0eqbj](https://cloud.githubusercontent.com/assets/6977319/17240448/25dbbe32-559e-11e6-8e1c-043cb6d0d6e5.png)