@halx99 2016-07-29
I will share to ```cpp-empty-test```