@minggo 2016-07-29
We should be more focus on iOS and Android. And the optimization should be:

* better performance on 64-bit devices
* equal or better performance on 32-bit devices

We can not accept an optimization that speed up on 64-bit devices but slow down 32-bit devices. As you know, most Android devices are 32-bit.