@halx99 2016-07-29
My test not use macro, I will use macro and test again, and share my test to ```cpp-empty-test```